Make CalicoEtcdClient cert optional for searcher (#18)
When introducing cert for calico etcd client keys it was initially
thought that adding it here doesn't break anything as nothing uses it.
This was wrong thinking. Adding it to searcher actually takes it into
use everywhere and initially this was fixed by creating corresponding
certconfig in kubernetesd and cluster-operator. Now there's still
problem with old clusters that don't have this so therefore it should be
made optional for now.